refactor(lean): migrate prover from semantic-kernel to Agent Framework
Split 3594-line monolith into prover/ package (12 modules):
- 4 specialized agents (Search, Tactic, Critic, Coordinator) with
  per-agent tool classes and distinct instructions
- WorkflowBuilder graph with conditional routing via Case/Default
- VerifyExecutor (non-LLM) for direct Lean compilation checks
- OpenAIChatClient for z.ai endpoint (api.z.ai, GLM-5.1)
- AutonomousProver (single-agent loop) for fast iteration
- multi_agent_proof.py reduced to thin CLI (~256 lines)

Delete: SimpleAgent, ProofAgentGroupChat, SingleAgentProver,
all semantic_kernel imports.
