+++ MyIA.AI.Notebooks/SymbolicAI/Lean/agent_tests/prover/__init__.py
@@ -1,0 +1,16 @@
+"""Multi-Agent Lean 4 Proof System -- Microsoft Agent Framework."""
+
+from .provers import MultiAgentSorryProver, AutonomousProver
+from .config import PROVIDERS, DEMOS
+from .trace import TraceLogger
+from .state import ProofState, SorryContext
+
+__all__ = [
+    "MultiAgentSorryProver",
+    "AutonomousProver",
+    "PROVIDERS",
+    "DEMOS",
+    "TraceLogger",
+    "ProofState",
+    "SorryContext",
+]
